refactor: remove old ot-ui search component
