--- src/components/Search/DropdownIndicator.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import SearchIcon from '@material-ui/icons/Search';
-
-const DropdownIndicator = ({ inputRef, innerProps }) => {
-  return <SearchIcon ref={inputRef} {...innerProps} color="inherit" />;
-};
-
-export default DropdownIndicator;
