--- src/components/_Search/SearchQuery.gql
@@ -1,33 +1,0 @@
-query SearchQuery($queryString: String!) {
-  search(queryString: $queryString) {
-    totalGenes
-    totalVariants
-    totalStudies
-    genes {
-      id
-      symbol
-      chromosome
-      start
-      end
-    }
-    variants {
-      id
-      rsId
-      chromosome
-      position
-      refAllele
-      altAllele
-    }
-    studies {
-      studyId
-      traitReported
-      pmid
-      pubAuthor
-      pubDate
-      pubJournal
-      nInitial
-      numAssocLoci
-      hasSumstats
-    }
-  }
-}
